refactor : render함수 가독성 개선
