+++ src/components/FilterSelector/CategorySelector/index.js
@@ -1,6 +1,6 @@
 import { FOOD_CATEGORY } from "../../../constants/foodCategory";
 import storage from "../../../domain/storage.ts";
-import renderFilteredRestaurants from "../../../renderFilteredRestaurant";
+import renderFilteredRestaurants from "../../../ui/renderFilteredRestaurant.js";
 import createKeyValuePair from "../../../utils/createKeyValuePair";
 import Select from "../../common/Select";
